refactor(ao-cli): split main.rs into cli/commands/session modules
Extract clap definitions to cli/args.rs, shared wiring (AutoScm, plugin
selection, project resolution, printing, local issues, spawn helpers,
browser, lifecycle notifier setup) under cli/, and one file per command
area under commands/. Session restore/attach live under session/.

Deduplicate notifier registration between watch and dashboard via
cli/lifecycle_wiring.rs. Keep main.rs as tracing init plus command
dispatch; move unit tests to src/tests.rs.

Behavior and CLI surface are unchanged. Verified with cargo test -p ao-cli,
cargo test --workspace, cargo clippy --workspace --all-targets.

Refs: #50
Made-with: Cursor
