+++ crates/ao-cli/src/cli/mod.rs
@@ -1,0 +1,12 @@
+//! Shared CLI helpers: arguments, plugins, printing, local issues, etc.
+
+pub mod agent_config;
+pub mod args;
+pub mod auto_scm;
+pub mod browser;
+pub mod lifecycle_wiring;
+pub mod local_issue;
+pub mod plugins;
+pub mod printing;
+pub mod project;
+pub mod spawn_helpers;
