+++ crates/ao-cli/src/commands/mod.rs
@@ -1,0 +1,13 @@
+//! Subcommand implementations (`ao-rs <cmd>`).
+
+pub mod cleanup;
+pub mod dashboard;
+pub mod doctor;
+pub mod kill;
+pub mod pr;
+pub mod review_check;
+pub mod send;
+pub mod spawn;
+pub mod start;
+pub mod status;
+pub mod watch;
