+++ crates/ao-cli/src/session/mod.rs
@@ -1,0 +1,4 @@
+//! `ao-rs session *` subcommands.
+
+pub mod attach;
+pub mod restore;
